fix(sprint-2b): cache lock + skip overflow + error propagation (Gemini #66)
Four MEDIUM findings from Gemini on PR #66:

1. evictExpired held the write lock while scanning the whole map, which
   stalls readers on a large cache. Switch to two-phase: read-lock to
   collect expired keys, write-lock only the deletions, with a
   re-check between locks so a refresh in the gap doesn't get blown
   away by mistake.

2. + 3. Skip multiplication was int64((page-1) * limit), which on
   32-bit GOARCH performs the multiply in int first and can wrap.
   int64(page-1) * int64(limit) does the widening up front.

4. /overview, /richlist, /blocksizes were dropping the error from
   GetOrCompute. The current fn bodies happen to return nil err, but
   that's brittle — return 500 explicitly so a future change can't
   accidentally serve 200 OK with a null body.

+++ backendAPI/db/transaction.go
@@ -111,7 +111,7 @@ func ReturnAllInternalTransactionsByAddress(address string, page, limit int) ([]
 	opts := options.Find().
 		SetProjection(projection).
 		SetSort(primitive.D{{Key: "blockTimestamp", Value: -1}}).
-		SetSkip(int64((page - 1) * limit)).
+		SetSkip(int64(page-1) * int64(limit)).
 		SetLimit(int64(limit))
 
 	results, err := configs.InternalTransactionByAddressCollection.Find(ctx, filter, opts)
@@ -181,7 +181,7 @@ func ReturnAllTransactionsByAddress(address string, page, limit int) ([]models.T
 	opts := options.Find().
 		SetProjection(projection).
 		SetSort(primitive.D{{Key: "timeStamp", Value: -1}}).
-		SetSkip(int64((page - 1) * limit)).
+		SetSkip(int64(page-1) * int64(limit)).
 		SetLimit(int64(limit))
 
 	results, err := configs.TransactionByAddressCollection.Find(ctx, filter, opts)
